+++ ManagerApplication/manager_application.cpp
@@ -27,7 +27,7 @@ int main()
 
 				"º±≈√: ";
 
-			int8_t function = 0;
+			int32_t function = 0;
 			switch (inputPositiveInteger(function))
 			{
 			case FUNCTION_CANCEL:
