+++ ManagerApplication/price_manager.cpp
@@ -21,7 +21,7 @@ PriceManager::PriceManager(DBHelper &dbHelper) : dbHelper(dbHelper)
 
 			"º±≈√: ";
 		
-		int8_t function = 0;
+		int32_t function = 0;
 		switch (inputPositiveInteger(function))
 		{
 		case FUNCTION_CANCEL:
